+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.0.3]
 
-### Fixed
-- Updated composer
+### Changed
+- Support php 8
 
 ## [v0.0.2]